+++ src/commands/image/imageSource/containerRegistry/acr/AcrListStep.ts
@@ -3,20 +3,17 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { KnownActiveRevisionsMode } from "@azure/arm-appcontainers";
 import { type ContainerRegistryManagementClient, type Registry } from "@azure/arm-containerregistry";
 import { type ResourceGroup } from "@azure/arm-resources";
-import { LocationListStep, ResourceGroupListStep, getResourceGroupFromId, parseAzureResourceId, uiUtils } from "@microsoft/vscode-azext-azureutils";
+import { LocationListStep, ResourceGroupListStep, getResourceGroupFromId, uiUtils } from "@microsoft/vscode-azext-azureutils";
 import { AzureWizardPromptStep, nonNullProp, type AzureWizardExecuteStep, type IAzureQuickPickItem, type ISubscriptionActionContext, type IWizardOptions } from "@microsoft/vscode-azext-utils";
-import { acrDomain, noMatchingResources, noMatchingResourcesQp } from "../../../../../constants";
+import { noMatchingResources, noMatchingResourcesQp } from "../../../../../constants";
 import { createContainerRegistryManagementClient } from "../../../../../utils/azureClients";
-import { parseImageName } from "../../../../../utils/imageNameUtils";
 import { localize } from "../../../../../utils/localize";
-import { currentlyDeployedPickDescription, hasMatchingPickDescription } from "../../../../../utils/pickUtils";
 import { type ContainerAppCreateBaseContext } from "../../../../createContainerApp/ContainerAppCreateContext";
 import { type ManagedEnvironmentCreateContext } from "../../../../createManagedEnvironment/ManagedEnvironmentCreateContext";
 import { type ContainerRegistryImageSourceContext } from "../ContainerRegistryImageSourceContext";
-import { getLatestContainerAppImage } from "../getLatestContainerImage";
+import { AcrDefaultPickUpdateStrategy } from "./AcrDefaultPickUpdateStrategy";
 import { type CreateAcrContext } from "./createAcr/CreateAcrContext";
 import { RegistryCreateStep } from "./createAcr/RegistryCreateStep";
 import { RegistryNameStep } from "./createAcr/RegistryNameStep";
@@ -28,8 +25,10 @@ export interface AcrListStepOptions {
     pickUpdateStrategy?: AcrPickUpdateStrategy;
 }
 
+export type AcrPickItem = IAzureQuickPickItem<Registry>;
+
 export interface AcrPickUpdateStrategy {
-    updatePicks(context: ContainerRegistryImageSourceContext, picks: IAzureQuickPickItem<Registry>[]): void | Promise<void>;
+    updatePicks(context: ContainerRegistryImageSourceContext, picks: AcrPickItem[]): AcrPickItem[] | Promise<AcrPickItem[]>;
 }
 
 const acrCreatePick = {
@@ -70,17 +69,16 @@ export class AcrListStep<T extends ContainerRegistryImageSourceContext> extends 
     }
 
     private async getPicks(context: T): Promise<IAzureQuickPickItem<Registry | typeof noMatchingResources | undefined>[]> {
-        const registryPicks: IAzureQuickPickItem<Registry>[] = (await AcrListStep.getRegistries(context)).map(r => {
+        const registryPicks: AcrPickItem[] = (await AcrListStep.getRegistries(context)).map(r => {
             return {
                 label: nonNullProp(r, 'name'),
                 data: r,
             };
         });
-        await this.options.pickUpdateStrategy?.updatePicks(context, registryPicks);
 
-        const picks: IAzureQuickPickItem<Registry | typeof noMatchingResources | undefined>[] = registryPicks;
+        const picks: IAzureQuickPickItem<Registry | undefined>[] = await this.options.pickUpdateStrategy?.updatePicks(context, registryPicks) ?? [];
         if (!this.options?.suppressCreatePick) {
-            picks.push(acrCreatePick);
+            picks.unshift(acrCreatePick);
         }
         if (!picks.length) {
             return [noMatchingResourcesQp];
@@ -140,101 +138,5 @@ async function tryConfigureResourceGroupForRegistry(
     resourceCreationContext.resourceGroup = resourceGroups.find(rg => resourceGroupName && rg.name === resourceGroupName);
     if (!resourceCreationContext.resourceGroup) {
         promptSteps.push(new ResourceGroupListStep());
-    }
-}
-
-export class AcrDefaultPickUpdateStrategy implements AcrPickUpdateStrategy {
-    updatePicks(context: ContainerRegistryImageSourceContext, picks: IAzureQuickPickItem<Registry>[]): void {
-        const registriesByGroup: Record<string, Registry[]> = {};
-        for (const p of picks) {
-            const registry: Registry | undefined = p.data;
-            if (!registry.id) {
-                continue;
-            }
-
-            const { resourceGroup } = parseAzureResourceId(registry.id);
-            const registriesGroup: Registry[] = registriesByGroup[resourceGroup] ?? [];
-            registriesGroup.push(registry);
-            registriesByGroup[resourceGroup] = registriesGroup;
-        }
-
-        // Sort resource groups alphabetically; if matches selected resource group, prioritize to the top
-        const sortedResourceGroups: string[] = Object.keys(registriesByGroup).sort((a, b) => {
-            const lowA: string = a.toLocaleLowerCase();
-            const lowB: string = b.toLocaleLowerCase();
-
-            switch (true) {
-                // If the user already picked a resource group, sort those to the top
-                case a === context.resourceGroup?.name:
-                    return -1;
-                case b === context.resourceGroup?.name:
-                    return 1;
-
-                // Everything below handles normal alphabetical sorting
-                case lowA < lowB:
-                    return -1;
-                case lowB < lowA:
-                    return 1;
-                default:
-                    return 0;
-            }
-        });
-
-        // Check for a currently deployed registry
-        let currentRegistry: string | undefined;
-        let hasCurrentRegistry: boolean = false;
-        if (context.containerApp) {
-            const { registryDomain, registryName } = parseImageName(getLatestContainerAppImage(context.containerApp, context.containersIdx ?? 0));
-            if (context.containerApp.revisionsMode === KnownActiveRevisionsMode.Single && registryDomain === acrDomain) {
-                currentRegistry = registryName;
-
-                const crIndex: number = picks.findIndex((p) => !!currentRegistry && p.data.loginServer === currentRegistry);
-                hasCurrentRegistry = crIndex !== -1;
-            }
-        }
-
-        const reOrderedPicks: IAzureQuickPickItem<Registry>[] = [];
-        let hasSameRgRegistry: boolean = false;
-        context.telemetry.properties.sameRgAcrCount = '0';
-        for (const [i, rg] of sortedResourceGroups.entries()) {
-            const registriesGroup: Registry[] = registriesByGroup[rg];
-
-            // Same resource group would be sorted to the top of the list...
-            let maybeSameRg: string | undefined;
-            if (i === 0 && !hasCurrentRegistry && rg === context.resourceGroup?.name) {
-                maybeSameRg = localize('sameRg', 'Within Same Resource Group');
-                context.telemetry.properties.sameRgAcrCount = String(registriesGroup.length);
-                hasSameRgRegistry = true;
-            }
-
-            // ...any "Other" resource groups would come after
-            let maybeOtherRg: string | undefined;
-            if (i > 0 && hasSameRgRegistry) {
-                maybeOtherRg = localize('other', 'Other');
-            }
-
-            const groupedRegistries: IAzureQuickPickItem<Registry>[] = registriesGroup.map(r => {
-                const maybeDeployed: string = r.loginServer === currentRegistry ? ` ${currentlyDeployedPickDescription}` : '';
-                return {
-                    label: nonNullProp(r, 'name'),
-                    group: maybeSameRg || maybeOtherRg,
-                    description: maybeDeployed || rg,
-                    data: r,
-                };
-            });
-
-            picks.push(...groupedRegistries);
-        }
-
-        // If a currently deployed registry exists, bring it to the front of the list
-        if (hasCurrentRegistry) {
-            const cdIdx: number = picks.findIndex(p => hasMatchingPickDescription(p, currentlyDeployedPickDescription));
-            if (cdIdx !== -1) {
-                const currentlyDeployedPick: IAzureQuickPickItem<Registry> | undefined = picks.splice(cdIdx, 1)[0];
-                picks.unshift(currentlyDeployedPick);
-            }
-        }
-
-        picks = reOrderedPicks;
     }
 }
